+++ tests/server/gameguard.test.js
@@ -6,7 +6,7 @@ const chai = require('chai');
 const sinon = require('sinon');
 const fs = require('fs-extra');
 
-const GameGuard = require('../../server/src/index');
+const GameGuard = require('../../build/index');
 const MockClient = require('./MockClient/MockClient');
 
 let server;
@@ -42,30 +42,24 @@ beforeEach(() => fs.truncateSync(dbPath, 0));
  * Before any test is run, we create a basic http server and use it to run the game server on.
  */
 before(done => {
-
   server = http.createServer(() => console.log('Server started'));
 
   gameGuard = new GameGuard(server, options);
 
   server.listen(7575, () => done());
-
 });
 
 /**
  * After all tests are finished running, we destory the game server object and close the server.
  */
 after(done => {
-
   server.on('close', () => done());
 
   server.close(() => {
-
     server.unref();
 
     process.exit();
-
   });
-
 });
 
 /**
@@ -74,69 +68,53 @@ after(done => {
  * ============================================================================================== 
  */
 describe('Players', () => {
-
   describe('Connecting players', () => {
-
     it('should add the player to the Array of connected players when the player connects', done => {
-
       addMockClients(1);
 
       setTimeout(() => {
-
-        chai.expect(gameGuard.players._players.length).to.equal(1);
+        chai.expect(gameGuard.players.connected.length).to.equal(1);
 
         done();
-
       }, 1000);
-
     });
 
     it('should create the player object saving their id and ip', done => {
-
       addMockClients(1);
 
       setTimeout(() => {
-
-        chai.expect(gameGuard.players._players[0].ip).to.equal('123.456.0.1') && chai.expect(gameGuard.players._players[0].id).to.equal('1dfd0497-a2f3-43de-9a55-d965bdde9d70');
+        chai.expect(gameGuard.players.connected[0].ip).to.equal('123.456.0.1') && chai.expect(gameGuard.players.connected[0].id).to.equal('1dfd0497-a2f3-43de-9a55-d965bdde9d70');
 
         done();
-
       }, 1000);
-
     });
 
     it('should emit and event when the player object is created', done => {
-
       const spy = sinon.spy();
 
-      gameGuard.players.on('player-joined', spy);
+      gameGuard.players.on('player-connected', spy);
 
       addMockClients(1);
 
       setTimeout(() => {
-
         chai.expect(spy.calledOnce).to.be.true;
 
         gameGuard.players.removeAllListeners();
 
         done();
-
       }, 1000);
-
     });
-
   });
 
   describe('Messaging players', () => {
-
     it('should send a message to a player', done => {
-
-      gameGuard.players.on('player-joined', player => {
+      gameGuard.players.on('player-connected', player => {
 
         player.message('info', 'hello there!');
 
-        chai.expect(player._ws.messages[0]).to.equal(`{"type":"info","content":"hello there!"}`);
+        const message = JSON.parse(player._socket.messages[0]);
 
+        chai.expect(message.type).to.equal('info') && chai.expect(message.contents).to.equal('hello there!');
       });
 
       addMockClients(1);
@@ -144,151 +122,121 @@ describe('Players', () => {
       gameGuard.players.removeAllListeners();
 
       done();
-
     });
 
     it('should send a message to multiple players', function (done) {
-
       this.timeout(2000);
 
       let players = [];
 
-      gameGuard.players.on('player-joined', player => {
-
+      gameGuard.players.on('player-connected', player => {
         players.push(player);
 
         player.message('debug', 'x: 5');
-
       });
 
       addMockClients(2);
 
       setTimeout(() => {
+        chai.expect(players[0]._socket.messages[0]).to.equal('{"type":"debug","contents":"x: 5"}') &&
 
-        chai.expect(players[0]._ws.messages[0]).to.equal('{"type":"debug","content":"x: 5"}') &&
-
-          chai.expect(players[1]._ws.messages[0]).to.equal('{"type":"debug","content":"x: 5"}')
+          chai.expect(players[1]._socket.messages[0]).to.equal('{"type":"debug","contents":"x: 5"}')
 
         gameGuard.players.removeAllListeners();
 
         done();
-
       }, 1500);
-
     });
-
   });
 
   describe('Kicking and banning players', () => {
 
     it('should kick a player', done => {
-
       const spy = sinon.spy();
 
       gameGuard.players.on('player-kicked', spy);
 
-      gameGuard.players.on('player-joined', player => player.kick('for testing'));
+      gameGuard.players.on('player-connected', player => player.kick('for testing'));
 
       addMockClients(1);
 
       setTimeout(() => {
-
         chai.expect(spy.calledOnce).to.be.true;
 
         gameGuard.players.removeAllListeners();
 
         done();
-
       }, 1000);
-
     });
 
     it('should kick a player with a reason', done => {
-
       const spy = sinon.spy();
 
       gameGuard.players.on('player-kicked', spy);
 
-      gameGuard.players.on('player-joined', player => player.kick('for testing'));
+      gameGuard.players.on('player-connected', player => player.kick('for testing'));
 
       addMockClients(1);
 
       setTimeout(() => {
-
         chai.expect(spy.getCalls()[0].args[1]).to.equal('for testing');
 
         gameGuard.players.removeAllListeners();
 
         done();
-
       }, 1000);
-
     });
 
     it('should ban a player', done => {
-
       const spy = sinon.spy();
 
       gameGuard.players.on('player-banned', spy);
 
-      gameGuard.players.on('player-joined', player => player.ban('for testing'));
+      gameGuard.players.on('player-connected', player => player.ban('for testing'));
 
       addMockClients(1);
 
       setTimeout(() => {
-
         chai.expect(spy.calledOnce).to.be.true;
 
         gameGuard.players.removeAllListeners();
 
         done();
-
       }, 1000);
-
     });
 
     it('should ban a player with a reason', done => {
-
       const spy = sinon.spy();
 
       gameGuard.players.on('player-banned', spy);
 
-      gameGuard.players.on('player-joined', player => player.ban('for testing'));
+      gameGuard.players.on('player-connected', player => player.ban('for testing'));
 
       addMockClients(1);
 
       setTimeout(() => {
-
         chai.expect(spy.getCalls()[0].args[1]).to.equal('for testing');
 
         gameGuard.players.removeAllListeners();
 
         done();
-
       }, 1000);
-
     });
 
     it('should ban a player and add them to the banned players list', async () => {
-
-      gameGuard.players.on('player-joined', player => player.ban('for testing'));
+      gameGuard.players.on('player-connected', player => player.ban('for testing'));
 
       addMockClients();
 
-      const banned = await gameGuard._storage._banned();
+      const banned = await gameGuard._storage.banned();
 
       setTimeout(() => {
-
         chai.expect(banned[0].id).to.equal('1dfd0497-a2f3-43de-9a55-d965bdde9d70');
 
         gameGuard.players.removeAllListeners();
-
       }, 1500);
-
     });
-
   });
-
 });
 
 /**
@@ -297,30 +245,23 @@ describe('Players', () => {
  * ============================================================================================== 
  */
 describe('Rooms', () => {
-
   describe('Creating rooms', () => {
-
     it('should create a room with default capacity', () => {
-
       const room1 = gameGuard.rooms.create('room1');
 
-      chai.expect(gameGuard.rooms.rooms[0].name).to.equal('room1') && chai.expect(gameGuard.rooms.rooms[0].capacity).to.equal(Infinity) && chai.expect(room1.name).to.equal('room1');
-
+      chai.expect(gameGuard.rooms.created[0].name).to.equal('room1') && chai.expect(gameGuard.rooms.created[0].capacity).to.equal(Infinity) && chai.expect(room1.name).to.equal('room1');
     });
 
     it('should create a room with a capacity of 10', () => {
-
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       gameGuard.rooms.create('room1', 10);
 
-      chai.expect(gameGuard.rooms.rooms[0].name).to.equal('room1') && chai.expect(gameGuard.rooms.rooms[0].capacity).to.equal(10);
-
+      chai.expect(gameGuard.rooms.created[0].name).to.equal('room1') && chai.expect(gameGuard.rooms.created[0].capacity).to.equal(10);
     });
 
     it('should create a room and emit an event when the room is created', done => {
-
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       const spy = sinon.spy();
 
@@ -333,12 +274,10 @@ describe('Rooms', () => {
       gameGuard.rooms.removeAllListeners();
 
       done();
-
     });
 
     it('should create a room and emit an event when the room is created with the room object', done => {
-
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       const spy = sinon.spy();
 
@@ -351,40 +290,33 @@ describe('Rooms', () => {
       gameGuard.rooms.removeAllListeners();
 
       done();
-
     });
 
     it('should fail creating a new room with the same name as an existing room', done => {
-
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       gameGuard.rooms.create('room1');
 
       chai.expect(() => gameGuard.rooms.create('room1')).to.throw('A room already exists with the name provided');
 
       done();
-
     });
-
   });
 
   describe('Destroying rooms', () => {
-
     it('should destroy a room', () => {
 
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       gameGuard.rooms.create('room1');
 
       gameGuard.rooms.destroy('room1');
 
-      chai.expect(gameGuard.rooms.rooms.length).to.equal(0);
-
+      chai.expect(gameGuard.rooms.created.length).to.equal(0);
     });
 
     it('should destroy a room and emit an event', done => {
-
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       const spy = sinon.spy();
 
@@ -399,12 +331,10 @@ describe('Rooms', () => {
       gameGuard.rooms.removeAllListeners();
 
       done();
-
     });
 
     it('should destroy a room and emit an event with the room name', done => {
-
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       const spy = sinon.spy();
 
@@ -419,18 +349,14 @@ describe('Rooms', () => {
       gameGuard.rooms.removeAllListeners();
 
       done();
-
     });
-
   });
 
   describe('Adding and removing players from rooms', () => {
-
     it('should add a player to a room', done => {
-
       const room1 = gameGuard.rooms.create('room1');
 
-      gameGuard.players.on('player-joined', player => room1.add(player));
+      gameGuard.players.on('player-connected', player => room1.add(player));
 
       addMockClients();
 
@@ -441,103 +367,80 @@ describe('Rooms', () => {
         done();
 
       }, 1500);
-
     });
 
     it('should remove a player from a room', done => {
-
       const players = [];
 
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       const room1 = gameGuard.rooms.create('room1');
 
-      gameGuard.players.on('player-joined', player => {
-
+      gameGuard.players.on('player-connected', player => {
         room1.add(player);
 
         players.push(player);
-
       });
 
       addMockClients(2);
 
       setTimeout(() => {
-
         room1.remove(players[0]);
 
         chai.expect(room1._players.length).to.equal(1) && chai.expect(room1._players[0]._id).to.equal(ids[1]);
 
         done();
-
       }, 1500);
-
     });
 
     it('should remove all players from a room', done => {
-
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       const room1 = gameGuard.rooms.create('room1');
 
-      gameGuard.players.on('player-joined', player => room1.add(player));
+      gameGuard.players.on('player-connected', player => room1.add(player));
 
       addMockClients(2);
 
       setTimeout(() => {
-
         room1.clear();
 
         chai.expect(room1._players.length).to.equal(0);
 
         done();
-
       }, 1500);
-
     });
-
   });
 
   describe('Messaging players in rooms', () => {
-
     it('should broadcast a message to all of the players in the room', function (done) {
-
       this.timeout(5000);
 
       let players = [];
 
-      gameGuard.rooms._rooms = [];
+      gameGuard.rooms._created = [];
 
       const broadcastRoom = gameGuard.rooms.create('broadcastRoom');
 
-      gameGuard.players.on('player-joined', player => {
-
+      gameGuard.players.on('player-connected', player => {
         players.push(player);
 
         broadcastRoom.add(player);
-
       });
 
       addMockClients(2);
 
       setTimeout(() => {
-
         broadcastRoom.broadcast('info', 'Hello World!');
 
         setTimeout(() => {
-
-          chai.expect(players[0]._ws.messages[0]).to.equal('{"type":"info","content":"Hello World!"}') && chai.expect(players[1]._ws.messages[0]).to.equal('{"type":"info","content":"Hello World!"}');
+          chai.expect(players[0]._socket.messages[0]).to.equal('{"type":"info","contents":"Hello World!"}') && chai.expect(players[1]._socket.messages[0]).to.equal('{"type":"info","contents":"Hello World!"}');
 
           done();
-
         }, 3000);
-
       }, 1500);
-
     });
-
   });
-
 });
 
 /**
@@ -546,29 +449,21 @@ describe('Rooms', () => {
  * ============================================================================================== 
  */
 describe('Messaging all players in the server', () => {
-
   it('should broadcast a message to all of the players in the server', function (done) {
-
     this.timeout(5000);
 
     addMockClients(2);
 
     setTimeout(() => {
-
       gameGuard.system.broadcast('info', 'Hello World!');
 
       setTimeout(() => {
-
-        chai.expect(gameGuard.players._players[0]._ws.messages[0]).to.equal('{"type":"info","content":"Hello World!"}') && chai.expect(gameGuard.players._players[1]._ws.messages[0]).to.equal('{"type":"info","content":"Hello World!"}');
-
+        chai.expect(gameGuard.players.connected[0]._socket.messages[0]).to.equal('{"type":"info","content":"Hello World!"}') && chai.expect(gameGuard.players.connected[1]._socket.messages[0]).to.equal('{"type":"info","content":"Hello World!"}');
       }, 3000);
-
       done();
 
     }, 1500)
-
   });
-
 });
 
 /**
@@ -584,7 +479,7 @@ function addMockClients(count = 1) {
 
     gameGuard._socket._events.connection(mockClient, mockClient.req);
 
-    const playerJoined = JSON.stringify({ type: 'player-joined', content: mockClient.id });
+    const playerJoined = JSON.stringify({ type: 'player-connected', contents: mockClient.id });
 
     mockClient.message(playerJoined);