+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 0.4.3 / 2020-01-08
 ==================
+* [FEATURE] Moved the client-side component to its own package, see the README for more.
+* [FEATURE] Messages now require a message object instead of type and content.
+* [FEATURE] Changed the `player-joined` event to `player-connected`.
+* [FEATURE] Changed the `player-left` event to `player-disconnected`.
+* [FEATURE] Converted the project to typescript.
+* [FEATURE] Changed the `players` property of the players module to be `connected`.
+* [FEATURE] Changed the `rooms` property of the rooms module to `created`.
+* [MISC] Removed unnecessary dependencies and moved express to a be a dev dependency which it should have been all along.
+* [MISC] Updated README to reflect breaking changes.
+
+0.4.3 / 2020-01-08
+==================
 * [MISC] Made private methods that weren't really private public.
 * [MISC] Updated dependencies to their latest versions.